change name of class/method/file
