+++ 09.wc_object/wc/filecontent.rb
