+++ 09.wc_object/wc/text.rb
@@ -3,17 +3,15 @@
 
 module Wc
   class Text
-    attr_reader :text
-
     def initialize(text)
       @text = text
     end
 
     def text_detail
       {
-        line_count: text.lines.size,
-        word_count: text.split(' ').size,
-        byte_count: text.size
+        line_count: @text.lines.size,
+        word_count: @text.split(' ').size,
+        byte_count: @text.size
       }
     end
   end
