+++ 09.wc_object/wc/textinline.rb
