+++ 09.wc_object/wc/wc.rb
